cmd: refactor commands to partially live in Service::Cmd::x
