+++ lib/service/commands/list.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+module Service
+  module Commands
+    module List
+      module_function
+
+      def run
+        ServicesCli.list
+      end
+    end
+  end
+end
