+++ lib/service/commands/stop.rb
@@ -1,0 +1,14 @@
+# frozen_string_literal: true
+
+module Service
+  module Commands
+    module Stop
+      module_function
+
+      def run(target, verbose:)
+        ServicesCli.check(target) &&
+          ServicesCli.stop(target, verbose: verbose)
+      end
+    end
+  end
+end
